+++ packages/ricos-content/web/src/converters/draft/fromDraft/fromDraft.ts
@@ -185,7 +185,8 @@ export const fromDraft = (draftJSON: DraftContent, opts: FromDraftOptions = {}):
       default:
     }
 
-    const nodes = getTextNodes(block, entityMap, opts);
+    const isEmptyTextNode = node => node.type === Node_Type.TEXT && node.textData?.text === '';
+    const nodes = getTextNodes(block, entityMap, opts).filter(node => !isEmptyTextNode(node));
 
     if (!isEmpty(nodes)) {
       paragraphNode.nodes = nodes;